Change Structure and add FindWineBottle Method
