+++ backend/src/main/java/fr/anettehaferkorn/backend/Recommendations/controller/exception/ApiError.java
@@ -1,0 +1,21 @@
+package fr.anettehaferkorn.backend.Recommendations.controller.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+import java.time.LocalDateTime;
+
+@Data
+@AllArgsConstructor
+public class ApiError {
+
+    private String text;
+    private String exceptionMessage;
+    private LocalDateTime timestamp;
+
+    public ApiError(String text, String exceptionMessage) {
+        this.text = text;
+        this.exceptionMessage = exceptionMessage;
+        this.timestamp = LocalDateTime.now();
+    }
+}
