--- backend/src/main/java/fr/anettehaferkorn/backend/controller/exception/NoMatchingWineException.java
@@ -1,14 +1,0 @@
-package fr.anettehaferkorn.backend.controller.exception;
-
-import java.util.NoSuchElementException;
-
-public class NoMatchingWineException extends NoSuchElementException {
-
-    public NoMatchingWineException(String message) {
-        super(message);
-    }
-
-    public NoMatchingWineException(String message, Exception e) {
-        super(message, e);
-    }
-}
