--- backend/src/main/java/fr/anettehaferkorn/backend/model/WineQuery.java
@@ -1,17 +1,0 @@
-package fr.anettehaferkorn.backend.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class WineQuery {
-
-    private String occasion;
-    private String wineStyle;
-    private String region;
-    private String alcohol;
-    private String taste;
-}
