+++ backend/src/test/java/fr/anettehaferkorn/backend/Recommendations/service/filter/FilterByTasteTest.java
@@ -1,0 +1,75 @@
+package fr.anettehaferkorn.backend.Recommendations.service.filter;
+
+import fr.anettehaferkorn.backend.Recommendations.model.RecommendationDTO;
+import fr.anettehaferkorn.backend.Recommendations.model.WineQuery;
+import org.junit.jupiter.api.Test;
+
+import java.util.List;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class FilterByTasteTest {
+
+    @Test
+    void matchByTaste() {
+
+        WineQuery wineQuery=new WineQuery("dinner","red","France","light","sweet");
+
+        List<RecommendationDTO> mockedRecommendations=List.of(
+                new RecommendationDTO("AAAA","dinner","red","France","light","sweet","abcd",1),
+                new RecommendationDTO("CCCC","dinner","red","France","strong","sweet","abcd",2),
+                new RecommendationDTO("EEEE","dinner","red","Spain","light","dry","abcd",2)
+        );
+
+        List<RecommendationDTO>mockedMatches=List.of(
+                new RecommendationDTO("AAAA","dinner","red","France","light","sweet","abcd",2),
+                new RecommendationDTO("CCCC","dinner","red","France","strong","sweet","abcd",3),
+                new RecommendationDTO("EEEE","dinner","red","Spain","light","dry","abcd",2)
+        );
+
+        List<RecommendationDTO> actual=FilterByTaste.matchByTaste(wineQuery,mockedRecommendations);
+
+        assertEquals(mockedMatches,actual);
+
+    }
+
+    @Test
+    void matchByTaste_noMatch() {
+
+        WineQuery wineQuery=new WineQuery("dinner","red","Italy","strong","dry");
+
+        List<RecommendationDTO>mockedRecommendations=List.of(
+                new RecommendationDTO("AAAA","dinner","red","France","light","sweet","abcd",1),
+                new RecommendationDTO("CCCC","dinner","red","France","light","sweet","abcd",2),
+                new RecommendationDTO("EEEE","dinner","red","Spain","light","sweet","abcd",2)
+        );
+
+        List<RecommendationDTO> actual=FilterByTaste.matchByTaste(wineQuery,mockedRecommendations);
+
+        assertEquals(mockedRecommendations,actual);
+
+    }
+
+    @Test
+    void matchByAlcohol_idK() {
+
+        WineQuery wineQuery=new WineQuery("dinner","red","idK","idK","idK");
+
+        List<RecommendationDTO>mockedRecommendations=List.of(
+                new RecommendationDTO("AAAA","dinner","red","France","light","sweet","abcd",1),
+                new RecommendationDTO("CCCC","dinner","red","France","strong","sweet","abcd",2),
+                new RecommendationDTO("EEEE","dinner","red","Spain","light","sweet","abcd",2)
+        );
+        List<RecommendationDTO>mockedMatches=List.of(
+                new RecommendationDTO("AAAA","dinner","red","France","light","sweet","abcd",2),
+                new RecommendationDTO("CCCC","dinner","red","France","strong","sweet","abcd",3),
+                new RecommendationDTO("EEEE","dinner","red","Spain","light","sweet","abcd",3)
+        );
+
+
+        List<RecommendationDTO> actual=FilterByTaste.matchByTaste(wineQuery,mockedRecommendations);
+
+        assertEquals(mockedMatches,actual);
+
+    }
+}
